Update page for owner

+++ app/DataTables/ProjectDataTable.php
@@ -4,6 +4,7 @@ namespace App\DataTables;
 
 use Carbon\Carbon;
 use Collective\Html\FormFacade;
+use Illuminate\Support\Facades\Auth;
 use Yajra\DataTables\DataTables;
 
 class ProjectDataTable{
@@ -17,8 +18,11 @@ class ProjectDataTable{
                 return 'Harian';
             })
             ->editColumn('progress', function($project){
-                // return getProgress($project->progress);
-                return FormFacade::select('type', [0 => 'Proyek Baru', 1 => 'Pembelian Bahan', 2 => 'Pengerjaan', 3 => 'Pemasangan', 4 => 'Selesai'], $project->progress, ['id' => 'projectType', 'onchange' => 'updateStatusProyek(this.value, '. $project->id .')', 'class' => 'form-control']);
+                if(checkUserLevel() == 'admin'){
+                    return FormFacade::select('progress', [0 => 'Proyek Baru', 1 => 'Pembelian Bahan', 2 => 'Pengerjaan', 3 => 'Pemasangan', 4 => 'Selesai'], $project->progress, ['id' => 'projectType', 'onchange' => 'updateProgressProyek(this.value, '. $project->id .')', 'class' => 'form-control']);
+                }
+                
+                return getProgress($project->progress);
             })
             ->editColumn('start', function($project){
                 return Carbon::parse($project->start)->format('d-F-Y');
@@ -38,18 +42,26 @@ class ProjectDataTable{
             ->addColumn('worker_salary', function($project){
                 return formatPrice($project->worker_salary);
             })
+            ->editColumn('approved', function($project){
+                return isApproved($project);
+            })
             ->addColumn('action', function($project){
-                $deleteUrl = "'".route('projects.admin.delete', $project->id)."', 'projectDatatable'";
+                if(checkUserLevel() == 'admin'){
+                    $deleteUrl = "'".route('projects.admin.delete', $project->id)."', 'projectDatatable'";
+                    return  '<div class="btn-group">'.
+                        '<a href="'.route('projects.admin.show_rab', $project->id).'" class="btn btn-primary">
+                            <i class="fas fa-tasks"></i>
+                        </a>'.
+                        '<a href="'.route('projects.admin.edit',$project->id).'" class="btn btn-warning" ><i class="menu-icon fa fa-pencil-alt"></i></a>'.
+                        '<a href="#" onclick="deleteModel('. $deleteUrl .')" class="btn btn-danger" ><i class="menu-icon fa fa-trash"></i></a>'.
+                    '</div>';
+                }
+
                 return  '<div class="btn-group">'.
                     '<a href="'.route('projects.admin.show_rab', $project->id).'" class="btn btn-primary">
                         <i class="fas fa-tasks"></i>
-                    </a>'.
-                    '<a href="'.route('projects.admin.edit',$project->id).'" class="btn btn-warning" ><i class="menu-icon fa fa-pencil-alt"></i></a>'.
-                    '<a href="#" onclick="deleteModel('. $deleteUrl .')" class="btn btn-danger" ><i class="menu-icon fa fa-trash"></i></a>'.
-                '</div>';
-
-                
-            })->addIndexColumn()->rawColumns(['action'])->make(true);
+                    </a>';                
+            })->addIndexColumn()->rawColumns(['action', 'approved'])->make(true);
     }
 
 }
